Objecter: don't attempt to read from non-primary on EC pools
With BALANCE_READS or LOCALIZE_READS set, the client will hang if
the non-primary OSD is picked because the OSD will most likely drop
the op (or start waiting for peering that won't actually happen).

Refactor the code so that the replica read conditions don't need to
be repeated.  Apart from the missing replica pool check, the acting
set size was checked only in the LOCALIZE_READS case.

Fixes: https://tracker.ceph.com/issues/45793
Signed-off-by: Ilya Dryomov <idryomov@gmail.com>