mgr/dashboard: Fix CephFS's Directories not showing
This is a regretion introduced in fdd8ec8ca9e38bc0cf9a7b31f232b0322b46782a.

We no longer need to check if a tab is selected to load the content.

In that commit we removed the logic that updated the conditional variable,
but  missed removing its use.

Fixes: https://tracker.ceph.com/issues/45877

Signed-off-by: Tiago Melo <tmelo@suse.com>