cmake: restrict crush_objs in libcrimson-alienstore.a.
The problem this commit deals with is that alienized
set of some `src/common` has included all `crush_srcs`:

  ```
  set(crush_srcs
    builder.c
    mapper.c
    crush.c
    hash.c
    CrushWrapper.cc
    CrushCompiler.cc
    CrushTester.cc
    CrushLocation.cc)
  ```

while only `CrushLocation.cc` has been altered with
the `TOPNSPC` macro -- other files lack the namespace
differentiation. This leads to a conflict with
`libcrimson-common.a` which includes the same files:

  ```
  add_library(crimson-common STATIC
  ...
    ${PROJECT_SOURCE_DIR}/src/crush/builder.c
    ${PROJECT_SOURCE_DIR}/src/crush/mapper.c
    ${PROJECT_SOURCE_DIR}/src/crush/crush.c
    ${PROJECT_SOURCE_DIR}/src/crush/hash.c
    ${PROJECT_SOURCE_DIR}/src/crush/CrushWrapper.cc
    ${PROJECT_SOURCE_DIR}/src/crush/CrushCompiler.cc
    ${PROJECT_SOURCE_DIR}/src/crush/CrushTester.cc
    ${PROJECT_SOURCE_DIR}/src/crush/CrushLocation.cc
  ```

Another, potential problem is linking with a variant
built `WITH_ALIEN` setting.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>