common/config: extract get_conffile_paths() out
md_config_t::parse_config_files() performs blocked i/o, and it might be
detected by the stall detector of seastar, and it spills out warning
messages when it sees the hang.

so i plan to implement an asynchronized parse_config_files() for
crimson, and reuse the synchronized part from the blocking
implementation of parse_config_files(). this is the first step.

Signed-off-by: Kefu Chai <kchai@redhat.com>