common/HeartbeatMap: use relaxed order
use memory_order_relaxed instead of memory_order_seq_cst, as the
ordering of storing of `timeout` and `suicide_timeout` does not matter,
and we don't need to synchorinize the cache for the consistent memory
view of local core. what we need is but the atomic read/write operation
of these individual variables.

Signed-off-by: Kefu Chai <kchai@redhat.com>