build/op, jaeger: fix target link issues
since cmake does not allow target_link_libraries for object type target
for cmake 3.10, this caused failure in including jaeger dependencies.
To fix it as a workaround until the time we can move to cmake 3.12,
including the directory directory to top level cmake.

crimson: link jaeger with crimson-osd

crimson uses some of the legacy osd code like OpRequest.cc which
references jaeger hence it also needs to be linked with jaeger

Signed-off-by: Deepika Upadhyay <dupadhya@redhat.com>