librbd: encode version in UpdateFeatures payload
with the recent support for async rbd operations from pacific+
when an older client(non async support) goes on upgrade, and simultaneously interacts with a newer
client which expects the requests to be async, experiences hang; considering the
return code for request completion to be acknowledgement for async request,
which then keeps waiting for another acknowledgement of request completion.

The idea is to encode version in async supported client request, so that
This case can be handled by letting this should make older clients error out
while trying to decode message with unsupported version.

fixes: https://tracker.ceph.com/issues/49592

Signed-off-by: Deepika <dupadhya@redhat.com>