octopus: qa/upgrade: disable update_features test_notify with older client as lockowner
with the recent support for async rbd operations from pacific+ when an
older client(non async support) goes on upgrade, and simultaneously
interacts with a newer client which expects the requests to be async,
experiences hang; considering the return code for request completion to
be acknowledgement for async request, which then keeps waiting for
another acknowledgement of request completion.

this if happens should be a rare only when lockowner is an old client
and should be deferred if compatibility issues arises.

Signed-off-by: Deepika Upadhyay <dupadhya@redhat.com>